+++ src/test/java/com/realexpayments/remote/sdk/utils/XmlUtilsTest.java
@@ -54,6 +54,7 @@ import static com.realexpayments.remote.sdk.utils.SampleXmlValidationUtils.RELEA
 import static com.realexpayments.remote.sdk.utils.SampleXmlValidationUtils.RECEIPT_IN_PAYMENT_REQUEST_XML_PATH;
 import static com.realexpayments.remote.sdk.utils.SampleXmlValidationUtils.PAYMENT_OUT_PAYMENT_REQUEST_XML_PATH;
 import static com.realexpayments.remote.sdk.utils.SampleXmlValidationUtils.PAYER_NEW_PAYMENT_REQUEST_XML_PATH;
+import static com.realexpayments.remote.sdk.utils.SampleXmlValidationUtils.PAYER_EDIT_PAYMENT_REQUEST_XML_PATH;
 import static com.realexpayments.remote.sdk.utils.SampleXmlValidationUtils.REQUEST_HASH;
 import static com.realexpayments.remote.sdk.utils.SampleXmlValidationUtils.RESPONSE_HASH;
 import static com.realexpayments.remote.sdk.utils.SampleXmlValidationUtils.RESULT_SUCCESS;
@@ -90,6 +91,7 @@ import static com.realexpayments.remote.sdk.utils.SampleXmlValidationUtils.check
 import static com.realexpayments.remote.sdk.utils.SampleXmlValidationUtils.checkUnmarshalledReceiptInPaymentRequest;
 import static com.realexpayments.remote.sdk.utils.SampleXmlValidationUtils.checkUnmarshalledPaymentOutPaymentRequest;
 import static com.realexpayments.remote.sdk.utils.SampleXmlValidationUtils.checkUnmarshalledPayerNewPaymentRequest;
+import static com.realexpayments.remote.sdk.utils.SampleXmlValidationUtils.checkUnmarshalledPayerEditPaymentRequest;
 import static com.realexpayments.remote.sdk.utils.SampleXmlValidationUtils.checkUnmarshalledThreeDSecureEnrolledResponse;
 import static com.realexpayments.remote.sdk.utils.SampleXmlValidationUtils.checkUnmarshalledThreeDSecureSigResponse;
 import static com.realexpayments.remote.sdk.utils.SampleXmlValidationUtils.checkUnmarshalledVerifyEnrolledRequest;
@@ -852,6 +854,21 @@ public class XmlUtilsTest {
 		//Convert from XML back to PaymentRequest
 		PaymentRequest fromXmlRequest = new PaymentRequest().fromXml(source);
 		checkUnmarshalledPayerNewPaymentRequest(fromXmlRequest);
+
+	}
+
+	/**
+	 * Tests conversion of {@link PaymentRequest} from XML file for payer-new payment types.
+	 */
+	@Test
+	public void paymentRequestXmlFromFilePayerEditTest() {
+
+		File file = new File(this.getClass().getResource(PAYER_EDIT_PAYMENT_REQUEST_XML_PATH).getPath());
+		StreamSource source = new StreamSource(file);
+
+		//Convert from XML back to PaymentRequest
+		PaymentRequest fromXmlRequest = new PaymentRequest().fromXml(source);
+		checkUnmarshalledPayerEditPaymentRequest(fromXmlRequest);
 
 	}
 }